Move userdelete to userservice.
